Log overall response message when using PRDR
When using PRDR, the "C=" log message only contains the PRDR response
and the overall response message (that typically contains the remote
message reference) is not logged.

The logged response is already capable of containing multiple lines
in the event that there is a multi-line response, so this change
appends the overall response to the PRDR response with a newline
between them.

The "C=" log message will therefore contain both responses:
=> ... PRDR C="250 PRDR R=<postmaster@example.com> acceptance\\n250 id=1eiQ9X-0005Ra-I3 message accepted" QT=3s DT=0s
-> ... PRDR C="250 PRDR R=<postmaster@example.net> acceptance\\n250 id=1eiQ9X-0005Ra-I3 message accepted" QT=3s DT=0s